+++ README.md
@@ -1,17 +1,19 @@
 **This project is in a pre-release stage and it is subject to change without warning.**
 
-[![Build Status](https://travis-ci.org/magicDGS/jsr203-http.svg?branch=master)](https://travis-ci.org/magicDGS/jsr203-http)
-[![Build status](https://ci.appveyor.com/api/projects/status/oiccmnsq7vvgy8tx/branch/master?svg=true)](https://ci.appveyor.com/project/magicDGS/jsr203-http/branch/master)
-[![codecov](https://codecov.io/gh/magicDGS/jsr203-http/branch/master/graph/badge.svg)](https://codecov.io/gh/magicDGS/jsr203-http)
-[![Codacy Badge](https://api.codacy.com/project/badge/Grade/6af974bdd86a4e7c9d4f7a71995f4dae)](https://www.codacy.com/app/daniel-gomez-sanchez/jsr203-http?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=magicDGS/jsr203-http&amp;utm_campaign=Badge_Grade)
+[![Build Status](https://travis-ci.org/lbergelson/http-nio.svg?branch=master)](https://travis-ci.org/lbergelson/http-nio)
+[![codecov](https://codecov.io/gh/lbergelson/http-nio/branch/master/graph/badge.svg)](https://codecov.io/gh/lbergelson/http-nio)
 [![Language](http://img.shields.io/badge/language-java-brightgreen.svg)](https://www.java.com/)
 [![License (3-Clause BSD)](https://img.shields.io/badge/license-BSD%203--Clause-blue.svg)](https://opensource.org/licenses/BSD-3-Clause)
 
-# jsr203-http
+# http-nio
 
 HTTP/S FileSystem provider for Java NIO.2
 
+# Origins
+This project was originally developed by [Daniel Gómez-Sánchez](https://github.com/magicDGS) as [magicDGS/jsr203-http](https://github.com/magicDGS/jsr203-http).
+This fork renames the project to be more easily understandable.  It also changes the package names so that it can be published under our maven central group.
+  
 ## License
 
 Code licensed under the BSD License. See the
-[LICENSE.txt](https://github.com/magicDGS/jsr203-http/blob/master/LICENSE.txt).
+[LICENSE.txt](https://github.com/lbergelson/http-nio/blob/master/LICENSE.txt).
