+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'jsr203-http'
+rootProject.name = 'http-nio'
