+++ src/main/java/org/broadinstitute/http/nio/HttpFileSystemProviderSettings.java
@@ -1,4 +1,4 @@
-package org.magicdgs.http.jsr203;
+package org.broadinstitute.http.nio;
 
 import java.net.Proxy;
