+++ src/main/java/org/broadinstitute/http/nio/package-info.java
@@ -1,4 +1,4 @@
 /**
  * HTTP/S FileSystem provider for Java NIO.2
  */
-package org.magicdgs.http.jsr203;
+package org.broadinstitute.http.nio;
