+++ src/main/resources/META-INF/services/java.nio.file.spi.FileSystemProvider
@@ -1,2 +1,2 @@
-org.magicdgs.http.jsr203.HttpFileSystemProvider
-org.magicdgs.http.jsr203.HttpsFileSystemProvider
+org.broadinstitute.http.nio.HttpFileSystemProvider
+org.broadinstitute.http.nio.HttpsFileSystemProvider
