+++ src/test/java/org/broadinstitute/http/nio/HttpsFileSystemProviderUnitTest.java
@@ -1,4 +1,4 @@
-package org.magicdgs.http.jsr203;
+package org.broadinstitute.http.nio;
 
 import org.testng.Assert;
 import org.testng.annotations.Test;
@@ -6,10 +6,10 @@ import org.testng.annotations.Test;
 /**
  * @author Daniel Gomez-Sanchez (magicDGS)
  */
-public class HttpFileSystemProviderUnitTest extends BaseTest {
+public class HttpsFileSystemProviderUnitTest extends BaseTest {
 
     @Test
     public void testGetScheme() throws Exception {
-        Assert.assertEquals(new HttpFileSystemProvider().getScheme(), "http");
+        Assert.assertEquals(new HttpsFileSystemProvider().getScheme(), "https");
     }
 }
