+++ src/test/java/org/broadinstitute/http/nio/URLSeekableByteChannelUnitTest.java
@@ -1,4 +1,4 @@
-package org.magicdgs.http.jsr203;
+package org.broadinstitute.http.nio;
 
 import org.testng.Assert;
 import org.testng.annotations.DataProvider;
@@ -6,7 +6,6 @@ import org.testng.annotations.Test;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.net.MalformedURLException;
 import java.net.URL;
 import java.nio.ByteBuffer;
 import java.nio.channels.ClosedChannelException;
